requirements.txt
flask
flask_login
pyshark
zxcvbn
pycryptodome
requests_html
beautifulsoup4
urllib3==1.25.8